savings fix

+++ run.sh
@@ -1,7 +1,14 @@
-rm -rf build
-mkdir build
-cd build
-cmake ..
-make
-cd bin
-./QuarkEngine
+#!/bin/bash
+
+if [ "$1" = "nb" ]; then
+    cd build/bin || exit 1
+    ./QuarkEngine
+else
+    rm -rf build
+    mkdir build
+    cd build || exit 1
+    cmake ..
+    make
+    cd bin || exit 1
+    ./QuarkEngine
+fi
